RabeyaMuna__diff__id_122

+++ framework/py/flwr/common/differential_privacy.py
@@ -31,7 +31,10 @@ from flwr.common.logger import log
 
 def get_norm(input_arrays: NDArrays) -> float:
     """Compute the L2 norm of the flattened input."""
-    array_norms = [np.linalg.norm(array.flat) for array in input_arrays]
+    array_norms = [
+        np.linalg.norm(array.flat)  # type: ignore[attr-defined]
+        for array in input_arrays
+    ]
     # pylint: disable=consider-using-generator
     return float(np.sqrt(sum([norm**2 for norm in array_norms])))
 
@@ -39,7 +42,9 @@ def get_norm(input_arrays: NDArrays) -> float:
 def add_gaussian_noise_inplace(input_arrays: NDArrays, std_dev: float) -> None:
     """Add Gaussian noise to each element of the input arrays."""
     for array in input_arrays:
-        noise = np.random.normal(0, std_dev, array.shape).astype(array.dtype)
+        noise = np.random.normal(0, std_dev, array.shape).astype(  # type: ignore[attr-defined]
+            array.dtype  # type: ignore[attr-defined]
+        )
         array += noise
